pyproject.toml
[project]
name = "fastapi-demo-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi-filter>=2.0.1",
    "fastapi-pagination>=0.14.0",
    "fastapi[standard]>=0.116.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.43",
]
